extract dataclasses for residences2030
