+++ generatorcore/residences2030/__init__.py
@@ -1,406 +1,8 @@
-from dataclasses import dataclass, field
-
+import imp
 from .. import residences2018, business2018
 from ..inputs import Inputs
 from ..utils import div
-
-
-@dataclass
-class Vars0:
-    # Used by g
-    cost_wage: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_emplo_com: float = None  # type: ignore
-    demand_emplo_new: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_com: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_com: float = None  # type: ignore
-
-
-@dataclass
-class Vars1:
-    # Used by g_consult
-    cost_wage: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_emplo_com: float = None  # type: ignore
-    demand_emplo_new: float = None  # type: ignore
-    emplo_existing: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_com: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_com: float = None  # type: ignore
-    ratio_wage_to_emplo: float = None  # type: ignore
-
-
-@dataclass
-class Vars2:
-    # Used by p
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_biomass: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    demand_emethan: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_emplo_new: float = None  # type: ignore
-    demand_heatnet: float = None  # type: ignore
-    demand_heatpump: float = None  # type: ignore
-    demand_solarth: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_com: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_com: float = None  # type: ignore
-
-
-@dataclass
-class Vars3:
-    # Used by p_buildings_total
-    area_m2: float = None  # type: ignore
-    area_m2_nonrehab: float = None  # type: ignore
-    area_m2_rehab: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_mro: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_emplo_new: float = None  # type: ignore
-    demand_heat_nonrehab: float = None  # type: ignore
-    demand_heat_rehab: float = None  # type: ignore
-    emplo_existing: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    fec_factor_averaged: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_com: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_com: float = None  # type: ignore
-    number_of_buildings_rehab: float = None  # type: ignore
-    pct_nonrehab: float = None  # type: ignore
-    pct_of_wage: float = None  # type: ignore
-    pct_rehab: float = None  # type: ignore
-    rate_rehab_pa: float = None  # type: ignore
-    ratio_wage_to_emplo: float = None  # type: ignore
-
-
-@dataclass
-class Vars4:
-    # Used by p_buildings_until_1919, p_buildings_1919_1948, p_buildings_1949_1978, p_buildings_1979_1995, p_buildings_1996_2004
-    area_m2: float = None  # type: ignore
-    area_m2_nonrehab: float = None  # type: ignore
-    area_m2_rehab: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    demand_heat_nonrehab: float = None  # type: ignore
-    demand_heat_rehab: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    fec_factor_averaged: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_per_x: float = None  # type: ignore
-    number_of_buildings_rehab: float = None  # type: ignore
-    pct_nonrehab: float = None  # type: ignore
-    pct_rehab: float = None  # type: ignore
-    rate_rehab_pa: float = None  # type: ignore
-
-
-@dataclass
-class Vars5:
-    # Used by p_buildings_2005_2011, p_buildings_2011_today
-    area_m2_nonrehab: float = None  # type: ignore
-    area_m2_rehab: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    demand_heat_nonrehab: float = None  # type: ignore
-    demand_heat_rehab: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    fec_factor_averaged: float = None  # type: ignore
-    number_of_buildings_rehab: float = None  # type: ignore
-    pct_nonrehab: float = None  # type: ignore
-    pct_rehab: float = None  # type: ignore
-
-
-@dataclass
-class Vars6:
-    # Used by p_buildings_new
-    area_m2: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    fec_factor_averaged: float = None  # type: ignore
-    pct_x: float = None  # type: ignore
-
-
-@dataclass
-class Vars7:
-    # Used by p_buildings_area_m2_com
-    area_m2: float = None  # type: ignore
-    area_m2_rehab: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    fec_factor_averaged: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_com: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_com: float = None  # type: ignore
-    invest_per_x: float = None  # type: ignore
-    number_of_buildings_rehab: float = None  # type: ignore
-    pct_x: float = None  # type: ignore
-
-
-@dataclass
-class Vars8:
-    # Used by r
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_emplo_com: float = None  # type: ignore
-    demand_emplo_new: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_com: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_com: float = None  # type: ignore
-
-
-@dataclass
-class Vars9:
-    # Used by s
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_cost_energy: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_fuel: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_emplo_new: float = None  # type: ignore
-    demand_heat_nonrehab: float = None  # type: ignore
-    demand_heat_rehab: float = None  # type: ignore
-    emplo_existing: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_com: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_com: float = None  # type: ignore
-    pct_energy: float = None  # type: ignore
-
-
-@dataclass
-class Vars10:
-    # Used by s_fueloil, s_lpg, s_biomass, s_coal, s_petrol, s_heatnet
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_combustion_based_per_MWh: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_cost_energy: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_fuel: float = None  # type: ignore
-    cost_fuel_per_MWh: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    pct_energy: float = None  # type: ignore
-
-
-@dataclass
-class Vars11:
-    # Used by s_solarth
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_combustion_based_per_MWh: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    area_ha_available: float = None  # type: ignore
-    area_ha_available_pct_of_action: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_cost_energy: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_fuel: float = None  # type: ignore
-    cost_fuel_per_MWh: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_emplo_new: float = None  # type: ignore
-    emplo_existing: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    energy_installable: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_com: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_com: float = None  # type: ignore
-    invest_per_x: float = None  # type: ignore
-    pct_energy: float = None  # type: ignore
-    pct_of_wage: float = None  # type: ignore
-    power_to_be_installed_pct: float = None  # type: ignore
-    ratio_wage_to_emplo: float = None  # type: ignore
-
-
-@dataclass
-class Vars12:
-    # Used by s_heatpump
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_combustion_based_per_MWh: float = None  # type: ignore
-    CO2e_production_based: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_cost_energy: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_fuel: float = None  # type: ignore
-    cost_fuel_per_MWh: float = None  # type: ignore
-    cost_wage: float = None  # type: ignore
-    demand_emplo: float = None  # type: ignore
-    demand_emplo_new: float = None  # type: ignore
-    emplo_existing: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    full_load_hour: float = None  # type: ignore
-    invest: float = None  # type: ignore
-    invest_com: float = None  # type: ignore
-    invest_pa: float = None  # type: ignore
-    invest_pa_com: float = None  # type: ignore
-    invest_per_x: float = None  # type: ignore
-    pct_energy: float = None  # type: ignore
-    pct_of_wage: float = None  # type: ignore
-    power_installed: float = None  # type: ignore
-    power_to_be_installed: float = None  # type: ignore
-    ratio_wage_to_emplo: float = None  # type: ignore
-
-
-@dataclass
-class Vars13:
-    # Used by s_gas
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_combustion_based_per_MWh: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_cost_energy: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_fuel: float = None  # type: ignore
-    cost_fuel_per_MWh: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    pct_energy: float = None  # type: ignore
-
-
-@dataclass
-class Vars14:
-    # Used by s_elec_heating, s_elec
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_combustion_based_per_MWh: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_cost_energy: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    pct_energy: float = None  # type: ignore
-
-
-@dataclass
-class Vars15:
-    # Used by s_emethan
-    CO2e_combustion_based: float = None  # type: ignore
-    CO2e_combustion_based_per_MWh: float = None  # type: ignore
-    CO2e_total: float = None  # type: ignore
-    CO2e_total_2021_estimated: float = None  # type: ignore
-    change_CO2e_pct: float = None  # type: ignore
-    change_CO2e_t: float = None  # type: ignore
-    change_cost_energy: float = None  # type: ignore
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    cost_climate_saved: float = None  # type: ignore
-    cost_fuel: float = None  # type: ignore
-    energy: float = None  # type: ignore
-    pct_energy: float = None  # type: ignore
-
-
-@dataclass
-class Vars16:
-    # Used by p_elec_elcon, p_elec_heatpump
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    demand_change: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    energy: float = None  # type: ignore
-
-
-@dataclass
-class Vars17:
-    # Used by p_other
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    demand_electricity: float = None  # type: ignore
-    energy: float = None  # type: ignore
-
-
-@dataclass
-class Vars18:
-    # Used by p_vehicles
-    change_energy_MWh: float = None  # type: ignore
-    change_energy_pct: float = None  # type: ignore
-    demand_change: float = None  # type: ignore
-    energy: float = None  # type: ignore
-
-
-@dataclass
-class R30:
-    g: Vars0 = field(default_factory=Vars0)
-    g_consult: Vars1 = field(default_factory=Vars1)
-    p: Vars2 = field(default_factory=Vars2)
-    p_buildings_total: Vars3 = field(default_factory=Vars3)
-    p_buildings_until_1919: Vars4 = field(default_factory=Vars4)
-    p_buildings_1919_1948: Vars4 = field(default_factory=Vars4)
-    p_buildings_1949_1978: Vars4 = field(default_factory=Vars4)
-    p_buildings_1979_1995: Vars4 = field(default_factory=Vars4)
-    p_buildings_1996_2004: Vars4 = field(default_factory=Vars4)
-    p_buildings_2005_2011: Vars5 = field(default_factory=Vars5)
-    p_buildings_2011_today: Vars5 = field(default_factory=Vars5)
-    p_buildings_new: Vars6 = field(default_factory=Vars6)
-    p_buildings_area_m2_com: Vars7 = field(default_factory=Vars7)
-    r: Vars8 = field(default_factory=Vars8)
-    s: Vars9 = field(default_factory=Vars9)
-    s_fueloil: Vars10 = field(default_factory=Vars10)
-    s_lpg: Vars10 = field(default_factory=Vars10)
-    s_biomass: Vars10 = field(default_factory=Vars10)
-    s_coal: Vars10 = field(default_factory=Vars10)
-    s_petrol: Vars10 = field(default_factory=Vars10)
-    s_heatnet: Vars10 = field(default_factory=Vars10)
-    s_solarth: Vars11 = field(default_factory=Vars11)
-    s_heatpump: Vars12 = field(default_factory=Vars12)
-    s_gas: Vars13 = field(default_factory=Vars13)
-    s_elec_heating: Vars14 = field(default_factory=Vars14)
-    s_emethan: Vars15 = field(default_factory=Vars15)
-    s_elec: Vars14 = field(default_factory=Vars14)
-    p_elec_elcon: Vars16 = field(default_factory=Vars16)
-    p_elec_heatpump: Vars16 = field(default_factory=Vars16)
-    p_other: Vars17 = field(default_factory=Vars17)
-    p_vehicles: Vars18 = field(default_factory=Vars18)
+from .r30 import R30
 
 
 def calc(inputs: Inputs, *, r18: residences2018.R18, b18: business2018.B18) -> R30:
